DUNE/Manuevers/BasicSwarm: inline methods for optimization and struct constructors for better initialization.

+++ src/DUNE/Maneuvers/BasicSwarm.cpp
@@ -222,7 +222,7 @@ namespace DUNE
     }
 
     void
-    BasicSwarm::fromLocalCoordinates(double x, double y, float* lat, float* lon)
+    BasicSwarm::fromLocalCoordinates(double x, double y, double* lat, double* lon)
     {
       *lat = m_rlat;
       *lon = m_rlon;
